requirements.txt
# 🐳 Православный портал - Production Requirements
# Django Framework
Django==4.2.16
django-environ==0.11.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-extensions==3.2.3

# Database & Cache
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0

# Web Server & Production
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0

# Background Tasks
celery==5.5.3
django-celery-beat==2.7.0
django-celery-results==2.5.1

# File Handling & Media
Pillow==11.0.0
python-magic==0.4.27

# API & Integrations
requests==2.32.3
urllib3==2.3.0
certifi==2025.4.26

# Security & Authentication
cryptography==44.0.0
django-allauth==65.3.0
django-otp==1.5.4

# Forms & UI
django-widget-tweaks==1.5.0
django-bootstrap5==24.3

# Search & Analytics
django-haystack==3.3.0
whoosh==2.7.4

# Admin Enhancements
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-import-export==4.1.1

# Development & Testing (commented out for production)
# django-debug-toolbar==4.4.6
# pytest==8.3.4
# pytest-django==4.9.0
# coverage==7.6.9

# Monitoring & Logging
sentry-sdk==2.19.2
django-log-viewer==1.1.7

# Performance & Optimization
django-cachalot==2.7.0
django-compressor==4.5.1

# Internationalization
django-modeltranslation==0.19.8
django-parler==2.3

# Rich Text & Content
django-ckeditor==6.7.2
django-summernote==0.8.20.0
markdown==3.7

# Email & Notifications
django-anymail==12.0

# PWA & Modern Features
django-pwa==1.1.0
django-manifest-loader==1.2.0

# Utils & Helpers
python-slugify==8.0.4
python-dateutil==2.9.0.post0
pytz==2024.2
